[modified]State Machineの部分を別ファイルに記述
